requirements.txt
colorama
discord.py
anyio